Prototype hacky merged episode layout
Changed the layout of podcasts to merge episode and track metadata to show one element with one set of controls when there's one track for a podcast. This also informs the user when there are no recordings at all for a set.

It's pretty hacky and I'm still struggling to figure out where everything is :D

- Added an air date placeholder for the podcast cards
- changed the title type for podcast cards?